requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==3.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3


